chore: refactor project structure
