+++ src/Backgrounder/BackgrounderLane.ts
@@ -1,53 +1,11 @@
+import ImagePicker from 'src/main'
 import { v4 } from 'uuid'
 
-import { ImagePicker } from '../ImagePicker'
-
-export interface BackgrounderJob {
-  /**
-   * Internally used UUID for the job
-   */
-  id: string
-  /**
-   * The type of job, in a small string.
-   *
-   * This is used as a "unique" identifier for the job.
-   * If you run it in a lane with `unique: true`, only
-   * one job of this type will be in the queue at a time.
-   */
-  type: string
-  /**
-   * The action that will be run when the job is executed
-   */
-  action: () => void | Promise<void>
-  /**
-   * If true, the job will be run immediately if the lane is free
-   */
-  eager?: boolean
-}
-
-export type BackgrounderQueue = BackgrounderJob[]
-
-export interface BackgrounderLaneProps {
-  type: string
-  queue: BackgrounderQueue
-  /**
-   * The time to wait between jobs in milliseconds
-   */
-  sleep: number
-  /**
-   * If true, only one job of this type can be in the queue at a time
-   */
-  unique: boolean
-  /**
-   * Determines which job to keep when enqueuing a unique job
-   *
-   * 'first' keeps the existing job and ignores the new one
-   * 'last' keeps the new job and removes the existing one
-   *
-   * @default 'first'
-   */
-  uniqueKeep: 'first' | 'last'
-}
+import {
+  BackgrounderQueue,
+  BackgrounderLaneProps,
+  BackgrounderJob,
+} from './types'
 
 export class BackgrounderLane {
   private plugin: ImagePicker
@@ -58,7 +16,7 @@ export class BackgrounderLane {
   private queue: BackgrounderQueue
   private sleepTime: number = 0
 
-  log = (...args: any[]) => {
+  log = (...args: unknown[]) => {
     this.plugin.log(`Lane [${this.type}] -> `, ...args)
   }
 
@@ -150,50 +108,5 @@ export class BackgrounderLane {
     this.queue = []
     this.running = false
     this.log('Cleared queue')
-  }
-}
-
-/**
- * General purpose background job runner :)
- *
- * This is used mostly to alleviate the main thread from
- * doing too much work at once. Things like indexing,
- * image processing, or other long-running tasks can be
- * run in the background.
- *
- * It's a FIFO queue, so jobs are run in the order they
- * are enqueued.
- */
-export class Backgrounder {
-  public lanes: Record<string, BackgrounderLane> = {}
-
-  constructor(public plugin: ImagePicker) {}
-
-  log = (...args: any[]) => {
-    this.plugin.log('Backgrounder -> ', ...args)
-  }
-
-  createLane = (lane: Omit<BackgrounderLaneProps, 'queue'>) => {
-    this.lanes[lane.type] = new BackgrounderLane(this.plugin, lane)
-  }
-
-  deleteLane = (type: string) => {
-    delete this.lanes[type]
-  }
-
-  stop = (type: string) => {
-    if (this.lanes[type]) {
-      this.lanes[type].clear()
-      this.log('Stopped lane:', type)
-    } else {
-      this.log('Lane not found:', type)
-    }
-  }
-
-  stopAll = () => {
-    for (const lane of Object.values(this.lanes)) {
-      lane.clear()
-    }
-    this.log('Stopped all lanes')
   }
 }
